Arreglando los tamaños de texto - wip
